requirements.txt
beautifulsoup4==4.12.2
deap==1.3.3
Flask==2.3.2
Flask_Cors==3.0.10
linkedin_scraper==2.11.1
matplotlib==3.7.1
networkx==3.1
python-dotenv==1.0.0
numpy==1.24.3
pdfminer.six==20221105
Pillow==9.0.1
PyPDF2==3.0.1
pyppeteer==1.0.2
Requests==2.31.0
scholarly==1.7.11
scikit_learn==1.2.2
selenium==4.10.0
spacy==3.5.3
tqdm==4.65.0
Unidecode==1.3.6
